+++ variables.tf
@@ -4,31 +4,32 @@ Module variables
 
 variable "tags" {
   description = "A map of tags to add to all resources."
+  type        = "map"
   default     = {}
 }
 
 variable "foo" {
-  description = "An example variable to be consumed for tests and resources."
+  description = "An example variable to be modified within tests."
   default     = "bar"
 }
 
 variable "baz" {
-  description = "An example variable to be consumed for tests and resources."
-  default     = "qux"
+  description = "Another example variable but without a default."
+  type        = "string"
 }
 
-variable "region" {}
-
-variable "list_input_output" {
-  type    = "list"
-  default = ["1", "2", 3]
+variable "list_with_defaults" {
+  description = "A list variable with defaults provided."
+  type        = "list"
+  default     = ["1", "2", "3"]
 }
 
-variable "map_input_output" {
-  type = "map"
+variable "list_without_defaults" {
+  description = "A list variable with defaults provided."
+  type        = "list"
+}
 
-  default = {
-    foo = "bar"
-    baz = "bomb"
-  }
+variable "map_without_defaults" {
+  description = "An example of a map used to store default values."
+  type        = "map"
 }