+++ CHANGELOG.md
@@ -1,11 +1,23 @@
 # Change Log
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.1.0] - 2017-xx-xx
+## [0.2.0] - 2017-09-28
+
 ### Added
-* Initial release
+
+* Docs created on terraform best practices.
+* `terraform-doc` now generating documentation.
 
 ### Changed
+
+* Updated to latest versions of terraform and gems.
+
+## [0.1.0] - 2017-09-28
+
+### Added
+
+* Initial release